added to requestConfig.json added file RequestDistance.json

+++ server/src/main/java/com/tco/server/RequestDistance.java
@@ -6,7 +6,7 @@ public class RequestDistance extends RequestHeader {
     Map<String, String> place1;
     Map<String, String> place2;
     Double earthRadius;
-    Integer distance;
+    Long distance;
 
     @Override
     public void buildResponse() {
